lib/flow.h Revert bitfield back to uint64_t.
Using different types for the two bitfields did not work on MSVC, so
reverting back to "64-bit bool" :-)

Reported-by: Saurabh Shah <ssaurabh@vmware.com>

+++ lib/flow.h
@@ -383,7 +383,7 @@ BUILD_ASSERT_DECL(FLOW_U32S <= 63);
  */
 struct miniflow {
     uint64_t map:63;
-    uint8_t values_inline:1;
+    uint64_t values_inline:1;
     union {
         uint32_t *offline_values;
         uint32_t inline_values[MINI_N_INLINE]; /* Minimum inline size. */
